requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
Django==2.1
django-celery-beat==1.4.0
django-crispy-forms==1.7.2
django-timezone-field==3.0
kombu==4.5.0
mysqlclient==1.3.13
Pillow==5.2.0
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2018.5
six==1.12.0
vine==1.3.0
